# Enhance benchmarking