@FilipAndersson245 2024-03-30
If someone in core could help me test this out it would be appreciated 👏