@FilipAndersson245 2024-03-30
I changed the default comparason metric from `median` to `fastest` based on [this](https://nikolaivazquez.com/blog/divan/#robust-benchmarking-paper)