@FilipAndersson245 2024-03-31
Im trying to rewrite this PR and instead using the main branch of divan and instead try to parse down the graphical output from divan.