@FilipAndersson245 2024-04-01
> I checked it, and it seems we can't publish a crate with unpublished dependencies, so the git link won't work.
> 
> It is a bit unfortunate that divan development has slowed down. Last commit was 2 months ago and divan's PR implementing file output hasn't yet received any maintainer comments despite being open since January. We can roll with your parsing script, but if divan ends up being unmaintained, we might need to switch back to criterion or something that's at least somewhat maintained.

I have a new script that parses the raw output, in the oven, probably try to get it in this PR, so we don`t have to use a PR