@FilipAndersson245 2024-04-01
@fdncred @kubouch I have tried out a new approach that relies on parsing the visual printout of divan, I think its decent, but may break if they change their format.

# The new markdown display format
|Name|Old|New|Relative|Note|
|-|-|-|-|-|
| load_standard_lib|8ms 502µs|8ms 502µs|1| |
| decoding_benchmarks|||||
|&ensp; json_decode|||||
|&ensp;&ensp; (100, 5)|406µs 200ns|406µs 200ns|1| |
|&ensp;&ensp; (10000, 15)|157ms 100µs|157ms 100µs|1| |
|&ensp; msgpack_decode|||||
|&ensp;&ensp; (100, 5)|179µs 100ns|179µs 100ns|1| |
|&ensp;&ensp; (10000, 15)|63ms 720µs|63ms 720µs|1| |
| encoding_benchmarks|||||
|&ensp; json_encode|||||
|&ensp;&ensp; (100, 5)|87µs 280ns|87µs 280ns|1| |
|&ensp;&ensp; (10000, 15)|21ms 50µs|21ms 50µs|1| |
|&ensp; msgpack_encode|||||
|&ensp;&ensp; (100, 5)|35µs 400ns|35µs 400ns|1| |
|&ensp;&ensp; (10000, 15)|9ms 212µs|9ms 212µs|1| |
| eval_benchmarks|||||
|&ensp; eval_default_config|2ms 216µs|2ms 216µs|1| |
|&ensp; eval_default_env|347µs 400ns|347µs 400ns|1| |
| eval_commands|||||
|&ensp; each|||||
|&ensp;&ensp; 1|51µs|51µs|1| |
|&ensp;&ensp; 5|267µs 400ns|267µs 400ns|1| |
|&ensp;&ensp; 10|489µs 400ns|489µs 400ns|1| |
|&ensp;&ensp; 100|5ms 339µs|5ms 339µs|1| |
|&ensp;&ensp; 1000|53ms 560µs|53ms 560µs|1| |
|&ensp; for_range|||||
|&ensp;&ensp; 1|51µs 320ns|51µs 320ns|1| |
|&ensp;&ensp; 5|266µs|266µs|1| |
|&ensp;&ensp; 10|532µs 100ns|532µs 100ns|1| |
|&ensp;&ensp; 100|5ms 247µs|5ms 247µs|1| |
|&ensp;&ensp; 1000|53ms 150µs|53ms 150µs|1| |
|&ensp; interleave|||||
|&ensp;&ensp; 100|420µs 400ns|420µs 400ns|1| |
|&ensp;&ensp; 1000|3ms 943µs|3ms 943µs|1| |
|&ensp;&ensp; 10000|38ms 730µs|38ms 730µs|1| |
|&ensp; interleave_with_ctrlc|||||
|&ensp;&ensp; 100|421µs 100ns|421µs 100ns|1| |
|&ensp;&ensp; 1000|3ms 784µs|3ms 784µs|1| |
|&ensp;&ensp; 10000|40ms 650µs|40ms 650µs|1| |
|&ensp; par_each_1t|||||
|&ensp;&ensp; 1|69µs 340ns|69µs 340ns|1| |
|&ensp;&ensp; 5|285µs 400ns|285µs 400ns|1| |
|&ensp;&ensp; 10|517µs 900ns|517µs 900ns|1| |
|&ensp;&ensp; 100|5ms 405µs|5ms 405µs|1| |
|&ensp;&ensp; 1000|53ms 300µs|53ms 300µs|1| |
|&ensp; par_each_2t|||||
|&ensp;&ensp; 1|77µs 750ns|77µs 750ns|1| |
|&ensp;&ensp; 5|189µs 200ns|189µs 200ns|1| |
|&ensp;&ensp; 10|304µs 700ns|304µs 700ns|1| |
|&ensp;&ensp; 100|2ms 785µs|2ms 785µs|1| |
|&ensp;&ensp; 1000|26ms 750µs|26ms 750µs|1| |
| parser_benchmarks|||||
|&ensp; parse_default_config_file|1ms 397µs|1ms 397µs|1| |
|&ensp; parse_default_env_file|247µs 100ns|247µs 100ns|1| |
| record|||||
|&ensp; create|||||
|&ensp;&ensp; 1|20µs 560ns|20µs 560ns|1| |
|&ensp;&ensp; 10|31µs 240ns|31µs 240ns|1| |
|&ensp;&ensp; 100|138µs 300ns|138µs 300ns|1| |
|&ensp;&ensp; 1000|1ms 229µs|1ms 229µs|1| |
|&ensp; flat_access|||||
|&ensp;&ensp; 1|11µs 210ns|11µs 210ns|1| |
|&ensp;&ensp; 10|11µs 890ns|11µs 890ns|1| |
|&ensp;&ensp; 100|19µs 100ns|19µs 100ns|1| |
|&ensp;&ensp; 1000|99µs 590ns|99µs 590ns|1| |
|&ensp; nest_access|||||
|&ensp;&ensp; 1|11µs 640ns|11µs 640ns|1| |
|&ensp;&ensp; 2|11µs 950ns|11µs 950ns|1| |
|&ensp;&ensp; 4|13µs 820ns|13µs 820ns|1| |
|&ensp;&ensp; 8|18µs 470ns|18µs 470ns|1| |
|&ensp;&ensp; 16|26µs 980ns|26µs 980ns|1| |
|&ensp;&ensp; 32|57µs 960ns|57µs 960ns|1| |
|&ensp;&ensp; 64|178µs 600ns|178µs 600ns|1| |
|&ensp;&ensp; 128|678µs 300ns|678µs 300ns|1| |
| table|||||
|&ensp; create|||||
|&ensp;&ensp; 1|23µs 710ns|23µs 710ns|1| |
|&ensp;&ensp; 10|35µs 870ns|35µs 870ns|1| |
|&ensp;&ensp; 100|156µs 900ns|156µs 900ns|1| |
|&ensp;&ensp; 1000|1ms 368µs|1ms 368µs|1| |
|&ensp; get|||||
|&ensp;&ensp; 1|16µs 340ns|16µs 340ns|1| |
|&ensp;&ensp; 10|19µs 580ns|19µs 580ns|1| |
|&ensp;&ensp; 100|50µs 90ns|50µs 90ns|1| |
|&ensp;&ensp; 1000|335µs 500ns|335µs 500ns|1| |
|&ensp; select|||||
|&ensp;&ensp; 1|15µs 480ns|15µs 480ns|1| |
|&ensp;&ensp; 10|22µs 30ns|22µs 30ns|1| |
|&ensp;&ensp; 100|79µs 50ns|79µs 50ns|1| |
|&ensp;&ensp; 1000|646µs 300ns|646µs 300ns|1| |
||||||
 
# Nushell record look
![image](https://github.com/nushell/nushell/assets/17986183/d66cd254-d13b-466a-8202-dac5dbddb7bb)