@kubouch 2024-04-01
Would it be possible to modify the output to not contain the empty entries? It could be made into a tree structure, or added as a prefix, e.g., `encoding_benchmarks::json_decode::(100, 5)`, `encoding_benchmarks::json_decode::(10000, 5)`, etc.

Otherwise, I think we could roll with this and see how it goes with divan's development in the future.