@FilipAndersson245 2024-04-01
I think this works fine, i am not to sure how to format the data correctly for that.