@kubouch 2024-04-01
It looks fine, but it's harder to do any processing with it. But we can keep it as-is, since it's still an improvement over having no parsing. The suggestion I made can be added later as another postprocessing step, for example.