@FilipAndersson245 2024-04-02
I am thinking on what you said @kubouch and maybe Criterion would be better, in the least part cause it so standardized in the community. And would most likely live on in some form in the longterm.