@kubouch 2024-04-02
Yeah, divan seems really cool, but is more uncertain than criterion. We could switch back to criterion and see what happens with divan.