@FilipAndersson245 2024-04-02
> Yeah, divan seems really cool, but is more uncertain than criterion. We could switch back to criterion and see what happens with divan.

My only problem is that the maintenance, status of Criterion is quite uncertain also, as the main contributor have not committed in almost 2 years. I am becoming inspired to write my own 😎.