@kubouch 2024-04-02
Haha, that's the way :-D. Maybe it would be worth checking with the divan maintainer and perhaps offering help with the maintenance instead of starting from scratch.

The main plus of criterion is being able to get some structured data out of it (e.g., JSON). I've been playing with ideas of how to set up continuous benchmark tracking, but it's quite essential to be able to reliably output structured data. If we could get the JSON PR landed to divan, we could just continue using divan as long as it works in the current form.