@FilipAndersson245 2024-04-05
I have create an alternative PR where i migrate the benchmarks back to criterion, but after the talk jon had, that was linked in the weekly meeting I have started to try out tango.rs and it seem very interesting, it works by compiling two versions of a program and comparing them. I may explore if that would be a better alternative. As I am not that sold on criterion  either as it's slow and not as robust as it gives the shine of.